Fix account sync giving role to bots

+++ backend/src/discord/apex-connect/apex-sync.service.ts
@@ -47,13 +47,17 @@ export class ApexSyncService {
 
     // Remove disconnected role from every user that has connected Apex Account
     for (const user of users.values()) {
-      if (usersWithConnectedApexAccount.some(apexAccount => apexAccount.user.discordId === user.id)) {
-        await this.discordService.removeRoleFromUser(user.id, disconnectRole.discordId);
-      }
-      // Else give user disconnected role
-      else if (!user.roles.cache.has(disconnectRole.discordId)) {
-        await this.discordService.addRoleToUser(user.id, disconnectRole.discordId);
-      }
+        // If user is bot skip him
+        if (user.user.bot)
+            continue;
+
+        if (usersWithConnectedApexAccount.some(apexAccount => apexAccount.user.discordId === user.id)) {
+            await this.discordService.removeRoleFromUser(user.id, disconnectRole.discordId);
+        }
+        // Else give user disconnected role
+        else if (!user.roles.cache.has(disconnectRole.discordId)) {
+            await this.discordService.addRoleToUser(user.id, disconnectRole.discordId);
+        }
     }
 
     return true;
